Add new event serializer strategy via a normalizer
There have been a number of recent issues with attempting to requeue domain events that then trigger serialization issues. More problems can be encountered if a message is requeued via the RabbitMQ console. This commit adds a new (de)normalizer plugin for the standard SF serializer and deprecates MessengerSerializer in this package.